refactor: ticket price 상수 위치 변경
