+++ src/main/java/lotto/domain/LottoMachine.java
@@ -4,11 +4,11 @@ import java.util.List;
 import java.util.Map;
 
 public class LottoMachine {
-    private static final int TICKET_PRICE = 1000;
+
     private final LottoTickets tickets;
 
     public LottoMachine(PurchaseAmount amount) {
-        this(new LottoTickets(amount.getTicketCount(TICKET_PRICE)));
+        this(new LottoTickets(amount));
     }
 
     public LottoMachine(LottoTickets tickets) {
@@ -24,19 +24,7 @@ public class LottoMachine {
     }
 
     public double getReturnRate(LottoTicket winningTicket) {
-        return (double) income(winningTicket) / cost();
-    }
-
-    private int income(LottoTicket winningTicket) {
-        return getRankStatistics(winningTicket)
-                .entrySet().stream()
-                .map(entry -> entry.getKey().getTotalPrize(entry.getValue()))
-                .reduce(Integer::sum)
-                .orElse(0);
-    }
-
-    private double cost() {
-        return Math.multiplyExact(TICKET_PRICE, tickets.getCount());
+        return tickets.getReturnRate(winningTicket);
     }
 
 }
